requirements.txt
absl-py==2.1.0
albumentations==1.4.6
annotated-types==0.6.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
contourpy==1.2.1
cupy-cuda11x==13.3.0
customtkinter==5.2.2
cycler==0.12.1
darkdetect==0.8.0
dask==2024.5.2
efficientnet==1.0.0
et-xmlfile==1.1.0
fastrlock==0.8.2
flatbuffers==24.3.25
fonttools==4.51.0
fsspec==2024.6.0
gast==0.4.0
graphviz==0.20.3
grpcio==1.63.0
h5py==3.11.0
idna==3.7
imageio==2.34.1
importlib_resources==6.4.0
joblib==1.4.2
keras==2.10.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
lazy_loader==0.4
libclang==18.1.1
locket==1.0.0
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.4
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
openpyxl==3.1.5
openslide-python==1.3.1
opt-einsum==3.3.0
pandas==2.2.2
partd==1.4.2
patchify==0.2.3
pillow==10.3.0
protobuf==3.19.6
pyasn1==0.6.0
pyasn1_modules==0.4.0
pydantic==2.7.1
pydantic_core==2.18.2
pydot==2.0.0
pyparsing==3.1.2
pytz==2024.1
pywin32==306
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==2.0.0
rsa==4.9
scikit-image==0.22.0
scikit-learn==1.4.2
scipy==1.13.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow-estimator==2.10.0
tensorflow-gpu==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.5.0
tifffile==2024.5.10
toolz==0.12.1
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
Werkzeug==3.0.3
wrapt==1.16.0
XlsxWriter==3.2.0